Added a own test project for environment variables testing.
