+++ src/Snapshooter.Environment.Tests/Helpers/SynchronExecutionFixture.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using Xunit;
+
+namespace Snapshooter.Environment.Tests.Helpers
+{
+    public static class CollectionFixtureNames
+    {
+        public const string SynchronExecutionFixture = "QueryCollectionFixture";
+    }
+}
